backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
bcrypt==3.2.0
pyjwt==2.1.0
gunicorn==20.1.0
stripe==2.58.0
python-jose==3.3.0
passlib==1.7.4 